move the notification model from the `services` directory to the `models` directory
