+++ lib/models/notification.dart
