+++ lib/screens/devices/deviceQrCode.dart
@@ -9,12 +9,12 @@ import 'package:flutter/material.dart';
 import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
 import 'package:nullpass/common.dart';
 import 'package:nullpass/models/device.dart';
+import 'package:nullpass/models/notification.dart' as np;
 import 'package:nullpass/models/qrData.dart';
 import 'package:nullpass/models/syncRegistration.dart';
 import 'package:nullpass/screens/appDrawer.dart';
 import 'package:nullpass/services/datastore.dart';
 import 'package:nullpass/services/logging.dart';
-import 'package:nullpass/services/notification.dart' as np;
 import 'package:openpgp/key_pair.dart';
 import 'package:openpgp/openpgp.dart';
 import 'package:qr_flutter/qr_flutter.dart' as qr;
