+++ lib/screens/devices/scanQrCode.dart
@@ -11,12 +11,12 @@ import 'package:flutter/services.dart';
 import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
 import 'package:nullpass/common.dart';
 import 'package:nullpass/models/device.dart';
+import 'package:nullpass/models/notification.dart' as np;
 import 'package:nullpass/models/qrData.dart';
 import 'package:nullpass/models/syncRegistration.dart';
 import 'package:nullpass/screens/appDrawer.dart';
 import 'package:nullpass/services/datastore.dart';
 import 'package:nullpass/services/logging.dart';
-import 'package:nullpass/services/notification.dart' as np;
 import 'package:nullpass/widgets.dart';
 import 'package:openpgp/key_options.dart';
 import 'package:openpgp/key_pair.dart';
